requirements.txt
Jinja2==3.0.2
croniter==1.0.15
fastapi==0.115.14
python-crontab==2.5.1
uvicorn==0.35.0
